requirements.txt
cantera>=3.1,<3.2
pyyaml>=6.0,<7
numpy>=1.26,<3
scipy>=1.11,<2
h5py>=3.10,<4
pytest>=8,<9
torch>=2.3,<3; platform_system == "Darwin" or platform_system == "Linux"
torch-geometric>=2.5,<3; platform_system == "Darwin" or platform_system == "Linux"
tgp>=0.0.1
optuna>=3.6,<5
